+++ packages/analytics_hub_mixpanel/test/src/resolver/mixpanel_event_resolver_test.dart
@@ -68,6 +68,8 @@ class _TestLogEvent extends LogEvent {
 
   @override
   List<EventProvider<LogEventResolver, LogEventOptions>> get providers => [
-        const EventProvider(MixpanelAnalyticsHubProviderKey(name: 'test')),
+        const EventProvider(
+          MixpanelAnalyticsHubProviderIdentifier(name: 'test'),
+        ),
       ];
 }